chore: rename types file
